@@ -3,14 +3,8 @@
   "description": "Guzzle Client for Wowza Server",
   "license": "MIT",
   "minimum-stability": "stable",
-  "authors": [
-    {
-      "name": "Jan Arnold",
-      "email": "jan.arnold@movingimage.com"
-    }
-  ],
   "require": {
-    "php": "^7.2|^8.0",
+    "php": "^7.4|^8.0",
     "guzzlehttp/guzzle": "^6.5",
     "symfony/config": "^5.2",
     "symfony/dependency-injection": "^5.2",
@@ -31,10 +25,5 @@
   },
   "config": {
     "bin-dir": "bin"
-  },
-  "extra": {
-    "branch-alias": {
-      "dev-master": "1.0.0-dev"
-    }
   }
 }
